cache atualizae e deletar recados

+++ src/app/features/recados/usecases/deletarRecado.useCase.ts
@@ -1,15 +1,31 @@
+import { CacheRepository } from "../../../shared/database/repositories/cache.repository";
 import { UsuarioRepository } from "../../usuario/repositories/usuario.repository";
 import { RecadoRepository } from "../repositories/recados.repository";
 
 export class DeletarRecadoUsecase {
-  constructor(private repository: RecadoRepository) {}
+  private cacheKey = "NOTE_DELETE";
+
+  constructor(
+    private repository: RecadoRepository,
+    private cacheRepository: CacheRepository
+  ) {}
 
   public async execute(recadoId: string, usuarioId: number): Promise<any> {
+    await this.cacheRepository.del(this.cacheKey);
+
+    const cache = await this.cacheRepository.get(this.cacheKey);
+
+    if (cache) {
+      return cache as any[];
+    }
+
     const result = await this.repository.delete(recadoId, usuarioId);
 
     if (!result) {
       return Error;
     }
+
+    await this.cacheRepository.set(this.cacheKey, result);
 
     return result;
   }
